// src/modules/ark/types.ts
export enum Profession {
  Caster = 'CASTER',
  Medic = 'MEDIC',
  Pioneer = 'PIONEER',
  Sniper = 'SNIPER',
  Special = 'SPECIAL',
  Support = 'SUPPORT',
  Tank = 'TANK',
  Token = 'TOKEN',
  Trap = 'TRAP',
  Warrior = 'WARRIOR',
}